Fixing James's comments:
my stupid curl effect loads before the image does.  Should put a blank
div for the image to load into so that no temporary effect.

rounded corners don't seem to work

that logo still bugs me...the png'd text has jagged edges black on
blue.  There isn't time right now, but I'd like to redo that with
proper image replacement for the text bits, so that the sucky
browsers get the image, but the better browsers get nice text with a
shadowed background.

I like the light blue menu, but when you hover it is light on gold and
is unreadable.  I arbitrarily tried #555 and it looked better
(master.css, line 170ish)

I like the gradient background.  looks like steel and asphalt and all
that.

I like the redundant pop up on the dropdown!

Proper link for calvad is right now lysithia.its.uci.edu/calvad

I shouldeventually be calvad.ctmlabs.net

but that apparently isn't on.